iOS 5.1 Compare
iOS 5.1 Compare

+++ Compare_Latest/Frameworks/AVFoundation/AVAssetTrackInternal.h
@@ -12,7 +12,7 @@
 {
     AVAssetTrackInspector *trackInspector;
     AVWeakReference *assetWeakReference;
-    int getTaggedCharacteristicsOncePerTrack;
+    long getTaggedCharacteristicsOncePerTrack;
     NSArray *taggedCharacteristics;
 }
