+++ Compare_Latest/Frameworks/AVFoundation/AVMetadataItemInternal.h
@@ -11,7 +11,7 @@
 @interface AVMetadataItemInternal : NSObject
 {
     struct OpaqueFigMetadataReader *reader;
-    int itemIndex;
+    long itemIndex;
     NSString *keySpace;
     id key;
     NSString *commonKey;
